# docs/swagger.yaml
basePath: /
definitions:
...
  dto.CreateProductRequest:
    properties:
      brand_id:
        type: integer
      category_id:
        type: integer
      description:
        type: string
      name:
        type: string
      slug:
        type: string
      status:
        type: string
    type: object
  dto.CreateProductVariantRequest:
    properties:
      discount_price:
        type: number
      price:
        type: number
      product_id:
        type: integer
      sku:
        type: string
    type: object
...
  dto.ProductResponse:
    properties:
      brand_id:
        type: integer
      category_id:
        type: integer
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      slug:
        type: string
      status:
        type: string
    type: object
  dto.ProductVariantResponse:
    properties:
      discount_price:
        type: number
      id:
        type: integer
      price:
        type: number
      product_id:
        type: integer
      sku:
        type: string
    type: object
...
  dto.UpdateProductRequest:
    properties:
      brand_id:
        type: integer
      category_id:
        type: integer
      description:
        type: string
      name:
        type: string
      slug:
        type: string
      status:
        type: string
    type: object
  dto.UpdateProductVariantRequest:
    properties:
      discount_price:
        type: number
      price:
        type: number
      sku:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: E-Commerce backend API built with Go, PostgreSQL, Redis, and Docker.
  title: E-Commerce API
  version: "1.0"
paths:
...
  /api/product-variants:
    post:
      consumes:
      - application/json
      description: Creates a new product variant
      parameters:
      - description: Product variant data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateProductVariantRequest'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.ProductVariantResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a new product variant
      tags:
      - ProductVariants
  /api/product-variants/{id}:
    delete:
      description: deletes a product variant by its ID.
      parameters:
      - description: Product Variant ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete product variant
      tags:
      - ProductVariants
    get:
      description: retrieves a product variant by its ID
      parameters:
      - description: Product Variant ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ProductVariantResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get a product variant by ID
      tags:
      - ProductVariants
    put:
      consumes:
      - application/json
      description: updates an existing product variant
      parameters:
      - description: Product variant ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated product variant data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateProductVariantRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ProductVariantResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: update  product variant
      tags:
      - ProductVariants
  /api/products:
    get:
      description: Retrieves all products.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.ProductResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get all products
      tags:
      - Products
    post:
      consumes:
      - application/json
      description: Creates a new product.
      parameters:
      - description: Product data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateProductRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.ProductResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a new product
      tags:
      - Products
  /api/products/{id}:
    delete:
      description: Soft deletes a product by its ID.
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete product
      tags:
      - Products
    get:
      description: Retrieves a product by its ID.
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ProductResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get product by ID
      tags:
      - Products
    put:
      consumes:
      - application/json
      description: Updates an existing product.
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated product data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ProductResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update product
      tags:
      - Products
  /api/products/{id}/variants:
    get:
      description: Retrieves all product variants belonging to a specific product.
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ProductVariantResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get all variants by product ID
      tags:
      - ProductVariants
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: Enter "Bearer " followed by your refresh token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"